% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_spike_by_firing_rate}
\alias{estimate_spike_by_firing_rate}
\title{Search for the right tuning parameter lambda to meet the target firing rate}
\usage{
estimate_spike_by_firing_rate(
  dat,
  decay_rate,
  target_firing_rate,
  lam_min = 1e-06,
  lam_max = 1,
  max_iters = 50,
  tolerance = 5
)
}
\arguments{
\item{dat}{A simdata object generated by simulate_ar1.}

\item{decay_rate}{Numeric; specified AR-1 decay rate \eqn{\gamma}, a 
number between 0 and 1 (non-inclusive).}

\item{target_firing_rate}{Numeric; a number between 0 to 1 indicating the average probability of firing in each timestep}

\item{lam_min}{Numeric; minimal lambda to consider}

\item{lam_max}{Numeric; maximal lambda to consider}

\item{max_iters}{Numeric; maximal iterations to search}

\item{tolerance}{Numeric; tolerance level for differences in firing rate}
}
\description{
Search for the right tuning parameter lambda to meet the target firing rate
}
\examples{
curr_sim <- simulate_ar1(n = 1000, gam = 0.95, poisMean = 0.01, sd = 0.1, seed = 1)
fit_spike <- estimate_spike_by_firing_rate(curr_sim, decay_rate = 0.95, 
target_firing_rate = 0.01, max_iters=10, tolerance=max(5))
}
