% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_spike_by_spike_number}
\alias{estimate_spike_by_spike_number}
\title{search for the right lambda to meet the targeting firing rate}
\usage{
estimate_spike_by_spike_number(
  dat,
  decay_rate,
  target_firing_rate,
  lam_min = 1e-06,
  lam_max = 1,
  max_iters = 50,
  tolerance = 5
)
}
\arguments{
\item{dat}{Numeric vector; observed data.}

\item{decay_rate}{Numeric; specified AR(1) decay rate \eqn{\gamma}, a 
number between 0 and 1 (non-inclusive).}

\item{target_firing_rate}{Numeric; a number between 0 to 1 
indicating the average probability of firing}

\item{lam_min}{Numeric; minimal lambda to consider}

\item{lam_max}{Numeric; maximal lambda to consider}

\item{max_iters}{Numeric; maximal iterations to search}

\item{tolerance}{Numeric; tolerance level for differences in firing rate}
}
\description{
search for the right lambda to meet the targeting firing rate
}
