% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.spike_inference}
\alias{plot.spike_inference}
\title{Plot the result of quantifying the uncertainty of spikes estimated via an L0 penalty}
\usage{
\method{plot}{spike_inference}(x, xlims = NULL, ...)
}
\arguments{
\item{x}{output from running spike_inference}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{to be passed to methods}
}
\description{
Plot the result of quantifying the uncertainty of spikes estimated via an L0 penalty
}
\examples{
gam <- 0.98
LAMBDA <- 0.7
sigma <- 0.3
n_length <- 1000
curr_sim <- simulate_ar1(n = n_length, gam = gam, poisMean = 0.01, sd = sigma, seed = 2)
curr_inference_spike <- spike_inference(dat = curr_sim$fl, decay_rate = gam,
 tuning_parameter = LAMBDA, window_size = 2, sig2 = sigma*sigma, return_ci = TRUE)
### The observed fluorescence is plotted in gray, along with two sets of 
### vertical ticks: each orange tick represents an estimated spike associated with 
### a positive increase in fluorescence, and blue ticks are the subset of spikes 
### with a selective p-value < 0.05.
plot(curr_inference_spike)
}
\references{
Chen YT, Jewell SW, Witten DM. (2021) Quantifying uncertainty in spikes estimated from calcium imaging data.
arXiv:2103.0781 [statME].
}
