% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.simdata}
\alias{print.simdata}
\title{Print simulated data}
\usage{
\method{print}{simdata}(x, ...)
}
\arguments{
\item{x}{simulated data}

\item{...}{arguments to be passed to methods}
}
\description{
Print simulated data
}
\examples{
sim <- simulate_ar1(n = 500, gam = 0.998, poisMean = 0.009, sd = 0.05, seed = 1)
print(sim)
}
