% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.spike_inference}
\alias{print.spike_inference}
\title{Print estimated spikes}
\usage{
\method{print}{spike_inference}(x, ...)
}
\arguments{
\item{x}{estimated spikes}

\item{...}{arguments to be passed to methods}
}
\description{
Print estimated spikes
}
\examples{
### Generate sample data
sim <- simulate_ar1(n = 500, gam = 0.998, poisMean = 0.01, sd = 0.05, seed = 1)
### Fit the spike
fit_spike <- spike_estimates(sim$fl, decay_rate = 0.998, tuning_parameter = 0.01)
print(fit_spike)
}
\keyword{internal}
