% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ar1.R
\name{simulate_ar1}
\alias{simulate_ar1}
\title{Simulate fluorescence trace based on simple AR(1) generative model}
\usage{
simulate_ar1(n, gam, poisMean, sd, seed)
}
\arguments{
\item{n}{Numeric; number of timesteps}

\item{gam}{Numeric; AR(1) decay rate}

\item{poisMean}{Numeric; mean for Poisson distributed spikes}

\item{sd}{Numeric; standard deviation}

\item{seed}{Numeric; random seed}
}
\value{
spikes, fluorescence, and calcium concentration
}
\description{
Simulate fluorescence trace based on simple AR(1) generative model
}
\details{
Simulate fluorescence trace based on simple AR(1) generative model

\deqn{y_t = c_t + \epsilon_t, \epsilon_t \sim N(0, \sigma^2)}

\deqn{c_t = gam * c_{t-1} + s_t}

\deqn{s_t ~ Pois(poisMean)}
}
\examples{
sim <- simulate_ar1(n = 500, gam = 0.998, poisMean = 0.009, sd = 0.05, seed = 1)
plot(sim)
}
