% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_inference.R
\name{spike_inference}
\alias{spike_inference}
\title{Estimation and inference for AR1 spike problem using an L0 penalty}
\usage{
spike_inference(
  dat,
  decay_rate,
  tuning_parameter,
  window_size,
  sig = NULL,
  return_conditioning_sets = FALSE,
  return_ci = TRUE,
  two_sided = FALSE,
  alpha = 0.05,
  mu = 0,
  lower_trunc = -Inf
)
}
\arguments{
\item{dat}{Numeric vector; observed data.}

\item{decay_rate}{Numeric; specified AR(1) decay rate \eqn{\gamma}, a 
number between 0 and 1 (non-inclusive).}

\item{tuning_parameter}{Numeric; tuning parameter \eqn{\lambda} for 
L0 spike estimation, a nonnegative number.}

\item{window_size}{Numeric; window size for fixed window hypothesis testing,
a nonnegative integer.}

\item{sig}{Numeric; noise variance for the observed data, a nonnegative number.
If unknown (NULL), sample variance of residuals is used instead.}

\item{return_ci}{Logical; if TRUE, the confidence interval for the change in calcium 
is computed and returned.}

\item{two_sided}{Logical; if TRUE, a 2-sided p-value is computed and returned.}

\item{alpha}{Numeric; significance level for the hypothesis test, 
a number between 0 and 1 (non-inclusive).}

\item{mu}{Numeric; parameter for the test  \eqn{\nu^{T}c\leq 0}.}

\item{return_conditioning_sets.}{Logical; Should the conditioning set S be returned?}
}
\value{
Returns a list with elements:

\code{spikes} the set of spikes,

\code{pvals}  p-values associated with each spike,

\code{LCB} lower confidence band for each spike,

\code{UCB} upper confidence band for each spike.
}
\description{
Estimation and inference for AR1 spike problem using an L0 penalty
}
\details{
Consider the AR(1) generative model \deqn{Y_t = c_t + \epsilon_t, \epsilon_t \sim N(0, \sigma^2)}
where \eqn{c_t = \gamma * c_{t-1} + z_t} and \eqn{z_t ~ Pois(poisMean)}. In words,
this says between spikes (when \eqn{z_t=0}), calcium decays exponentially at a known rate \eqn{\gamma\in(0,1)}, 
which is taken to be known. Further denote the locations of true spikes, \eqn{t:z_t\geq 0} as 
\eqn{0 = \tau_0 < \tau_1 < \ldots < \tau_K < \tau_{K+1} = T}.

Estimation:

This function first estimates spikes via L0 penalty based on
noisy observations \eqn{y_t, t = 1,  \ldots, T} by solving the optimization problem

\deqn{\hat{c}_1, \cdots, \hat{c}_T \in minimize{c_1,...,c_T\geq 0} 0.5 \sum_{t=1}^T ( y_t - c_t )^2 + \lambda \sum_t=2^T 
1(c_t != \gamma c_t-1) }.

Estimated spikes correspond to the time t such that estimated calcium does not decay exponentially, i.e.,
\eqn{\{\cdots,\hat{\tau}_j,\cdots\}  = \{t: \hat{c}_{t+1} -\gamma c_{t} \neq 0\} }.

See Jewell et al. (2019), Rigaill, G. (2015), 
and Maidstone, R. et al. (2017) for more information and discussion of
the detailed algorithm.


Inference:

For each estimated changepoint \eqn{\hat{\tau}_j}, we test the null
hypothesis that the calcium is decaying exponentially *near* \eqn{\hat{\tau}_j}.
In particular, near is defined based on a fixed window around
\eqn{\hat{\tau}_j} with left endpoint \deqn{tL = max(1, \hat{\tau}_j - window_size + 1),} and
right endpoint \deqn{tR = min(T, \hat{\tau}_j + window_size).}

We then test the null hypothesis \eqn{H_0: \sum_t v_t * \mu_t = 0}. 
for a linear contrast vector \eqn{v in R^T} defined as \eqn{v_t = 0}, if
\eqn{t < tL} or \eqn{t > tR}; 
\eqn{v_t = \gamma \cdot (\gamma^2-1)/(\gamma^{2(tL-\hat{\tau}_j)}-\gamma^2) \cdot \gamma^{t-\hat{\tau}_j}}, if
\eqn{tL \le t \le \hat{\tau}_j}; 
and \eqn{v_t = (\gamma^2-1)/(\gamma^{2(tR-\hat{\tau}_j)}-1) \cdot \gamma^{t-(\hat{\tau}_j}+1)}, if
\eqn{\hat{\tau}_j + 1 \le t \le tR}. 

Our framework allows us to efficiently compute p-values based on the 
test statistic \eqn{t(v) * y = \sum_t v_t \cdot y_t} for conditioning set \eqn{\{ \hat{\tau}_j in
M(y'(\phi)) \} }, where \eqn{\phi = \sum_t v_t * Y_t}. 

In Chen et al. (2020+), we show that the p-value corresponding
to the test \eqn{H0: \sum_t v_t * c_t = \mu} can be written as
\deqn{Pr(|\phi| \ge |t(v) * y| | \phi in S)} for a conditioning set S.

This function computes p-values for the following test statistic and
conditioning set S. In what follows, \eqn{y'(\phi)} is a perturbation of the
observed data y. See Proposition 1 of Chen et al. (2020+) for additional
details.

\deqn{Pr(\phi >= t(v) * y |  \hat{\tau}_j in M(y'(\phi))}

Since the observation \eqn{y_t} is normally distributed, \eqn{\phi | S} follows a truncated 
normal distribution, and we obtain the (one-sided or two-sided) p-value associated with  \eqn{ \hat{\tau}_j}
by computing the set S and find the CDF of the resulting truncated normal distribution.

By the duality of hypothesis testing and confidence interval, we can construct a \eqn{1-\alpha} confidence interval for
the parameter \eqn{t(v) * c = \sum_t v_t \cdot c_t}. This function computes the lower confidence band and upper confidence
band using a bisection method. See Proposition 8 of Chen et al. (2020+) for additional details.
}
\examples{

gam <- 0.98
LAMBDA <- 0.7
sigma <- 0.3
n_length <- 1000
curr_sim <- simulate_ar1(n = n_length, gam = gam, poisMean = 0.01, sd = sigma, seed = 2)
curr_fit_spike <- spike_estimates(dat = curr_sim$fl, decay_rate = gam, tuning_parameter = LAMBDA)
curr_inference_spike <- spike_inference(dat = curr_sim$fl, decay_rate = gam,
 tuning_parameter = LAMBDA, window_size = 2,
  sig = sigma*sigma, return_ci = TRUE)
}
\references{
Chen, Y., Jewell, S., and Witten, D. (2020+). Uncertainty quantification for
spikes estimated from calcium imaging data. Technical report.

Jewell, S. W., Hocking, T. D., Fearnhead, P., & Witten, D. M. (2019). 
Fast nonconvex deconvolution of calcium imaging data. Biostatistics. 

Jewell, S., Fearnhead, P., and Witten, D. (2019+). Testing for a change in
mean after changepoint detection. Technical report.

Maidstone, R., Hocking, T., Rigaill, G., & Fearnhead, P. (2017). On optimal
multiple changepoint algorithms for large data. Statistics and Computing,
27(2), 519-533.

Rigaill, G. (2015). A pruned dynamic programming algorithm to recover the
best segmentations with 1 to K_max change-points. Journal de la Societe
Francaise de Statistique, 156(4), 180-205.
}
