% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.spike_estimates}
\alias{summary.spike_estimates}
\title{Summarize the solution to an L0 segmentation problem}
\usage{
\method{summary}{spike_estimates}(object, ...)
}
\arguments{
\item{object}{output from running spike_estimates}

\item{...}{to be passed to methods}
}
\value{
A data frame with a column of estimated spikes
}
\description{
Summarize the solution to an L0 segmentation problem
}
\examples{
gam <- 0.98
LAMBDA <- 0.7
sigma <- 0.3
n_length <- 1000
curr_sim <- simulate_ar1(n = n_length, gam = gam, poisMean = 0.01, sd = sigma, seed = 2)
curr_fit_spike <- spike_estimates(dat = curr_sim$fl, decay_rate = gam, tuning_parameter = LAMBDA)
summary(curr_fit_spike)
}
