% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.spike_inference}
\alias{summary.spike_inference}
\title{Summarize the inference result for spikes estimated via an L0 inference problem}
\usage{
\method{summary}{spike_inference}(object, ...)
}
\arguments{
\item{object}{output from running spike_estimates}

\item{...}{to be passed to methods}
}
\value{
A data frame with a column of estimated spikes with associated p-values and confidence intervals.
}
\description{
Summarize the inference result for spikes estimated via an L0 inference problem
}
\examples{
gam <- 0.98
LAMBDA <- 0.7
sigma <- 0.3
n_length <- 1000
curr_sim <- simulate_ar1(n = n_length, gam = gam, poisMean = 0.01, sd = sigma, seed = 2)
curr_fit_spike <- spike_estimates(dat = curr_sim$fl, decay_rate = gam, tuning_parameter = LAMBDA)
curr_inference_spike <- spike_inference(dat = curr_sim$fl, decay_rate = gam,
tuning_parameter = LAMBDA, window_size = 2, sig2 = sigma*sigma, return_ci = TRUE)
summary(curr_inference_spike)
}
