% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_v}
\alias{construct_v}
\title{Generate the contrast vector for testing the presence of spikes}
\usage{
construct_v(n, thj, window_size, gam)
}
\arguments{
\item{n}{length of the vector}

\item{thj}{location of the spike of interest}

\item{window_size}{parameter h}

\item{gam}{AR-1 decaying parameter}
}
\description{
Generate the contrast vector for testing the presence of spikes
}
