% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.spike_estimates}
\alias{plot.spike_estimates}
\title{Plot the solution to an L0 segmentation problem}
\usage{
\method{plot}{spike_estimates}(x, xlims = NULL, ...)
}
\arguments{
\item{x}{output from running estimate_spikes}

\item{xlims}{optional parameter to specify the x-axis limits}

\item{...}{to be passed to methods}
}
\description{
Plot the solution to an L0 segmentation problem
}
