% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_inference.R
\name{spike_inference}
\alias{spike_inference}
\title{Estimation and inference for AR-1 spike problem using an L0 penalty}
\usage{
spike_inference(
  dat,
  decay_rate,
  tuning_parameter,
  window_size,
  sig2 = NULL,
  return_conditioning_sets = FALSE,
  return_ci = FALSE,
  two_sided = FALSE,
  alpha = 0.05,
  mu = 0
)
}
\arguments{
\item{dat}{Numeric vector; observed data.}

\item{decay_rate}{Numeric; specified AR-1 decay rate \eqn{\gamma}, a 
number between 0 and 1 (non-inclusive).}

\item{tuning_parameter}{Numeric; tuning parameter \eqn{\lambda} for 
L0 spike estimation, a non-negative number.}

\item{window_size}{Numeric; window size for fixed window hypothesis testing,
a non-negative integer.}

\item{return_conditioning_sets}{Logical; Should the conditioning set S be returned?}

\item{return_ci}{Logical; if TRUE, the confidence interval for the change in calcium 
is computed and returned.}

\item{two_sided}{Logical; if TRUE, a 2-sided p-value is computed and returned.}

\item{alpha}{Numeric; significance level for the hypothesis test, 
a number between 0 and 1 (non-inclusive).}

\item{mu}{Numeric; parameter for the test  \eqn{\nu^{T}c\leq mu}.}

\item{sig}{Numeric; noise variance for the observed data, a non-negative number.
If unknown (NULL), sample variance of residuals is used instead.}
}
\value{
Returns a list with elements:
\itemize{
\item \code{spikes} the set of spikes,
\item \code{pvals}  p-values associated with each spike,
\item \code{LCB} lower confidence band for each spike,
\item \code{UCB} upper confidence band for each spike.
}
}
\description{
Estimation and inference for AR-1 spike problem using an L0 penalty
}
\details{
Consider the AR-1 generative model \deqn{Y_t = c_t + \epsilon_t, \epsilon_t \sim N(0, \sigma^2),}
where \eqn{c_t = \gamma  c_{t-1} + z_t} and \eqn{z_t \sim Poisson(poisMean)}. In words,
this says between spikes (when \eqn{z_t=0}), calcium decays exponentially at a known rate \eqn{\gamma\in(0,1)}, 
which is taken to be known. Further denote the locations of true spikes, \eqn{\{t:z_t\geq 0\}} as 
\eqn{\{0 = \tau_0 < \tau_1 < \ldots < \tau_K < \tau_{K+1} = T\}}.

This function first estimates spikes via L0 penalty based on
noisy observations \eqn{y_t, t = 1,  \ldots, T} by solving the following optimization problem
\deqn{ minimize_{c_1,...,c_T\geq 0} \frac{1}{2} \sum_{t=1}^{T} ( y_t - c_t )^2 + \lambda \sum_{t=2}^{T} 
1(c_t \neq \gamma c_t-1).}
Estimated spikes correspond to the time t such that estimated calcium does not decay exponentially, i.e.,
\eqn{\{\cdots,\hat{\tau}_j,\cdots\}  = \{t: \hat{c}_{t+1} -\gamma c_{t} \neq 0\} }.

Now suppose that we want to test whether the calcium is exponentially decaying near an estimated spike
\eqn{\hat{\tau}_j}; or equivalently, the null hypothesis of the form \eqn{H_{0}:  \nu^T c = 0} versus 
\eqn{H_{1}:  \nu^T c> 0} for suitably chosen \eqn{\nu} (see Section 2 in Chen et al. (2021+) for details). 
This function computes the following p-value
\deqn{P(\nu^T Y \geq \nu^T y |  \hat{\tau}_j \in M(Y), \nu^T Y> 0, \Pi_\nu^\perp Y  =  \Pi_\nu^\perp y)}
where \eqn{M(Y)} is the set of spikes estimated from Y via the L0 method, \eqn{\Pi_\nu^\perp} is 
the orthogonal projection to the orthogonal complement of \eqn{\nu}. In particular, this p-value controls the 
selective Type I error (see Section 3 in  Chen et al. (2021+) for details). 

In addition, we implement a \eqn{1-\alpha} confidence interval for the parameter \eqn{ \nu^T c}, the increase
in calcium associated with an estimated spike \eqn{\hat{\tau}_j} (see Section 4 in Chen et al. (2021+) for details).
}
\examples{

gam <- 0.98
LAMBDA <- 0.7
sigma <- 0.3
n_length <- 1000
curr_sim <- simulate_ar1(n = n_length, gam = gam, poisMean = 0.01, sd = sigma, seed = 2)
curr_fit_spike <- spike_estimates(dat = curr_sim$fl, decay_rate = gam, tuning_parameter = LAMBDA)
curr_inference_spike <- spike_inference(dat = curr_sim$fl, decay_rate = gam,
 tuning_parameter = LAMBDA, window_size = 2, sig2 = sigma*sigma, return_ci = TRUE)
}
\references{
Chen YT, Jewell SW, Witten DM. (2021+) Quantifying uncertainty in spikes estimated from calcium imaging data.
 arXiv:2103.0781 [statME].
 
Jewell, S. W., Hocking, T. D., Fearnhead, P., & Witten, D. M. (2019). 
Fast nonconvex deconvolution of calcium imaging data. Biostatistics. 

Jewell, S., Fearnhead, P., and Witten, D. (2019+). Testing for a change in
mean after changepoint detection. Technical report.

Maidstone, R., Hocking, T., Rigaill, G., & Fearnhead, P. (2017). On optimal
multiple changepoint algorithms for large data. Statistics and Computing,
27(2), 519-533.

Rigaill, G. (2015). A pruned dynamic programming algorithm to recover the
best segmentations with 1 to K_max change-points. Journal de la Societe
Francaise de Statistique, 156(4), 180-205.
}
